Make hybrid search perf test that force merge index before queries.
